feat: Refactor home component to use filterable playlists
This commit refactors the home component to use the filterable playlists from the store instead of the featured playlists. It updates the HTML template and TypeScript file to reflect this change. By using the filterable playlists, the home component now displays a more comprehensive list of playlists that can be searched and filtered by the user.